--- packages/analytics/analytics-utilities/docs/interfaces/AnalyticsExploreRecord.md
@@ -1,41 +1,0 @@
-[@kong-ui-public/analytics-utilities](../analytics-utils.md) / AnalyticsExploreRecord
-
-# Interface: AnalyticsExploreRecord
-
-## Table of contents
-
-### Properties
-
-- [event](AnalyticsExploreRecord.md#event)
-- [timestamp](AnalyticsExploreRecord.md#timestamp)
-- [version](AnalyticsExploreRecord.md#version)
-
-## Properties
-
-### event
-
-• **event**: [`RecordEvent`](RecordEvent.md)
-
-#### Defined in
-
-[types/analytics-data.ts:8](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L8)
-
-___
-
-### timestamp
-
-• **timestamp**: `string`
-
-#### Defined in
-
-[types/analytics-data.ts:7](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L7)
-
-___
-
-### version
-
-• `Optional` **version**: `string`
-
-#### Defined in
-
-[types/analytics-data.ts:6](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L6)
