--- packages/analytics/analytics-utilities/docs/interfaces/AnalyticsExploreResult.md
@@ -1,36 +1,0 @@
-[@kong-ui-public/analytics-utilities](../analytics-utils.md) / AnalyticsExploreResult
-
-# Interface: AnalyticsExploreResult
-
-Analytics query data object
-
-## Table of contents
-
-### Properties
-
-- [meta](AnalyticsExploreResult.md#meta)
-- [records](AnalyticsExploreResult.md#records)
-
-## Properties
-
-### meta
-
-• **meta**: [`AnalyticsExploreMeta`](AnalyticsExploreMeta.md)
-
-Additional metadata about the query
-
-#### Defined in
-
-[types/analytics-data.ts:112](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L112)
-
-___
-
-### records
-
-• **records**: [`AnalyticsExploreRecord`](AnalyticsExploreRecord.md)[]
-
-Array of records
-
-#### Defined in
-
-[types/analytics-data.ts:108](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L108)
