--- packages/analytics/analytics-utilities/docs/interfaces/AnalyticsExploreV2Result.md
@@ -1,36 +1,0 @@
-[@kong-ui-public/analytics-utilities](../analytics-utils.md) / AnalyticsExploreV2Result
-
-# Interface: AnalyticsExploreV2Result
-
-Analytics exploreV2 query data object
-
-## Table of contents
-
-### Properties
-
-- [meta](AnalyticsExploreV2Result.md#meta)
-- [records](AnalyticsExploreV2Result.md#records)
-
-## Properties
-
-### meta
-
-• **meta**: [`AnalyticsExploreV2Meta`](AnalyticsExploreV2Meta.md)
-
-Additional exploreV2 metadata about the query
-
-#### Defined in
-
-[types/analytics-data.ts:126](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L126)
-
-___
-
-### records
-
-• **records**: [`AnalyticsExploreRecord`](AnalyticsExploreRecord.md)[]
-
-Array of records
-
-#### Defined in
-
-[types/analytics-data.ts:122](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L122)
