--- packages/analytics/analytics-utilities/docs/interfaces/AnalyticsExploreV3Result.md
@@ -1,36 +1,0 @@
-[@kong-ui-public/analytics-utilities](../analytics-utils.md) / AnalyticsExploreV3Result
-
-# Interface: AnalyticsExploreV3Result
-
-Analytics exploreV3 query data object
-
-## Table of contents
-
-### Properties
-
-- [meta](AnalyticsExploreV3Result.md#meta)
-- [records](AnalyticsExploreV3Result.md#records)
-
-## Properties
-
-### meta
-
-• **meta**: [`AnalyticsExploreV3Meta`](AnalyticsExploreV3Meta.md)
-
-Additional exploreV3 metadata about the query
-
-#### Defined in
-
-[types/analytics-data.ts:140](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L140)
-
-___
-
-### records
-
-• **records**: [`AnalyticsExploreRecord`](AnalyticsExploreRecord.md)[]
-
-Array of records
-
-#### Defined in
-
-[types/analytics-data.ts:136](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/analytics-data.ts#L136)
