--- packages/analytics/analytics-utilities/docs/interfaces/DruidGranularity.md
@@ -1,63 +1,0 @@
-[@kong-ui-public/analytics-utilities](../analytics-utils.md) / DruidGranularity
-
-# Interface: DruidGranularity
-
-## Table of contents
-
-### Properties
-
-- [duration](DruidGranularity.md#duration)
-- [origin](DruidGranularity.md#origin)
-- [period](DruidGranularity.md#period)
-- [timeZone](DruidGranularity.md#timezone)
-- [type](DruidGranularity.md#type)
-
-## Properties
-
-### duration
-
-• `Optional` **duration**: `number`
-
-#### Defined in
-
-[types/druid-granularity.ts:3](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/druid-granularity.ts#L3)
-
-___
-
-### origin
-
-• `Optional` **origin**: `string` \| `number`
-
-#### Defined in
-
-[types/druid-granularity.ts:6](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/druid-granularity.ts#L6)
-
-___
-
-### period
-
-• `Optional` **period**: `string`
-
-#### Defined in
-
-[types/druid-granularity.ts:4](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/druid-granularity.ts#L4)
-
-___
-
-### timeZone
-
-• `Optional` **timeZone**: `string`
-
-#### Defined in
-
-[types/druid-granularity.ts:5](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/druid-granularity.ts#L5)
-
-___
-
-### type
-
-• **type**: `string`
-
-#### Defined in
-
-[types/druid-granularity.ts:2](https://github.com/Kong/public-ui-components/blob/main/packages/analytics/analytics-utilities/src/types/druid-granularity.ts#L2)
